python_code/api/requirements.txt
pandas
python-dotenv
openai
runpod
pinecone
httpx
python-telegram-bot